the number of datapoints is an integer

+++ kois/lightcurve.h
@@ -1,7 +1,7 @@
 #ifndef _LIGHTCURVE_H_
 #define _LIGHTCURVE_H_
 
-void lightcurve (double n, double *t, int K, double texp, int np,
+void lightcurve (int n, double *t, int K, double texp, int np,
                  double *periods, double *epochs, double *durations,
                  double *rors, double *impacts, int nbins, double *r,
                  double *ir, double *f);
